requirements.txt
hidapi==0.11.0.post2
pyserial==3.5
